+++ src/ginkgoRunTestCodeLens.ts
@@ -5,30 +5,25 @@ import * as vscode from 'vscode';
 import { getTestSpecs } from './ginkgoSpecProvider';
 
 export class GinkgoRunTestCodeLensProvider implements vscode.CodeLensProvider {
-    public provideCodeLenses(document: vscode.TextDocument, token: vscode.CancellationToken): Thenable<vscode.CodeLens[]> {
+    public async provideCodeLenses(document: vscode.TextDocument, token: vscode.CancellationToken): Promise<vscode.CodeLens[]> {
         const ginkgoConfig = vscode.workspace.getConfiguration('ginkgolens');
+        const codeLenses: vscode.CodeLens[] = [];
 
-        return Promise.all([
-            this.getCodeLensesForFile(document),
-            this.getCodeLensesForIts(document)
-        ]).then(res => {
-            const codeLenses: vscode.CodeLens[] = [];
+        if (ginkgoConfig.get<boolean>('includeFile'))
+            codeLenses.push(... await this.getCodeLensesForFile(document));
 
-            if (ginkgoConfig.get<boolean>('includeFile'))
-                codeLenses.push(...res[0])
-            if (ginkgoConfig.get<boolean>('includeIts'))
-                codeLenses.push(...res[1]);
+        if (ginkgoConfig.get<boolean>('includeIts'))
+            codeLenses.push(... await this.getCodeLensesForIts(document));
 
-            return codeLenses;
-        });
+        return codeLenses;
     }
 
-    private getCodeLensesForFile(document: vscode.TextDocument): Thenable<vscode.CodeLens[]> {
+    private async getCodeLensesForFile(document: vscode.TextDocument): Promise<vscode.CodeLens[]> {
         if (document.fileName.endsWith('_suite_test.go')) {
-            return Promise.resolve([]);
+            return [];
         }
 
-        return Promise.resolve([
+        return [
             new vscode.CodeLens(
                 new vscode.Range(0, 0, 0, 0),
                 {
@@ -37,21 +32,19 @@ export class GinkgoRunTestCodeLensProvider implements vscode.CodeLensProvider {
                     arguments: [{ path: document.uri.fsPath }]
                 }
             )
-        ]);
+        ];
     }
 
-    private getCodeLensesForIts(document: vscode.TextDocument): Thenable<vscode.CodeLens[]> {
-        return getTestSpecs(document)
-            .then(testSpecs =>
-                testSpecs.map(spec =>
-                    new vscode.CodeLens(
-                        spec.location.range,
-                        {
-                            title: 'run test',
-                            command: 'ginkgo.test.focus',
-                            arguments: [{ testFocus: spec.fullSpecString }]
-                        }
-                    )
+    private async getCodeLensesForIts(document: vscode.TextDocument): Promise<vscode.CodeLens[]> {
+        return (await getTestSpecs(document))
+            .map(spec =>
+                new vscode.CodeLens(
+                    spec.location.range,
+                    {
+                        title: 'run test',
+                        command: 'ginkgo.test.focus',
+                        arguments: [{ testFocus: spec.fullSpecString }]
+                    }
                 )
             );
     }