+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 All notable changes to the "ginkgo-lens" extension will be documented in this file.
 
+## [Unreleased]
+- Switch from Promises to async/await paradigm
+
 ## [1.0.0]
 - Initial release